EventsGateway moved to shared
