+++ src/builds/builds.module.ts
@@ -4,11 +4,11 @@ import { BuildsController } from './builds.controller';
 import { UsersModule } from '../users/users.module';
 import { PrismaService } from '../prisma/prisma.service';
 import { TestRunsModule } from '../test-runs/test-runs.module';
-import { EventsGateway } from '../events/events.gateway';
+import { SharedModule } from '../shared/shared.module';
 
 @Module({
-  imports: [UsersModule, TestRunsModule],
-  providers: [BuildsService, PrismaService, EventsGateway],
+  imports: [SharedModule, UsersModule, TestRunsModule],
+  providers: [BuildsService, PrismaService],
   controllers: [BuildsController],
   exports: [BuildsService],
 })
