+++ src/builds/builds.service.ts
@@ -3,7 +3,7 @@ import { CreateBuildDto } from './dto/build-create.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { Build, Project } from '@prisma/client';
 import { TestRunsService } from '../test-runs/test-runs.service';
-import { EventsGateway } from '../events/events.gateway';
+import { EventsGateway } from '../shared/events/events.gateway';
 import { BuildDto } from './dto/build.dto';
 import uuidAPIKey from 'uuid-apikey';
