+++ src/shared/shared.module.ts
@@ -1,10 +1,11 @@
 import { Global, Module } from '@nestjs/common';
 import { StaticService } from './static/static.service';
+import { EventsGateway } from 'src/shared/events/events.gateway';
 
 @Global()
 @Module({
-    providers: [StaticService],
-    exports: [StaticService],
+    providers: [StaticService, EventsGateway],
+    exports: [StaticService, EventsGateway],
     imports: [],
     controllers: [],
 })
